+++ src/queries/script.ts
@@ -1,72 +1,15 @@
-import { DEFAULT_TAGS_RETRO, SCRIPT_CREATION_PAYMENT_TAGS, TAG_NAMES } from './constants';
-import { IContractEdge, IEdge, ITagFilter } from './interface';
-import { findByTags } from './queries';
+import { FairScript } from '../classes';
+import { DEFAULT_TAGS_RETRO, SCRIPT_CREATION_PAYMENT_TAGS, TAG_NAMES } from '../utils/constants';
+import { ITagFilter, IContractEdge, IEdge, listFilterParams } from '../types';
 import {
+  logger,
+  findByTags,
   filterByUniqueScriptTxId,
   filterPreviousVersions,
   findTag,
   getTxOwner,
   isFakeDeleted,
-  logger,
-} from './utils';
-
-type listScriptsParam = string | IContractEdge | IEdge;
-
-/**
- * @description Class to wrap a Fair Protocol Script tx with easy to access proeprties
- * @property {string} owner - Owner of the script
- * @property {string} name - Name of the script
- * @property {string} txid - Transaction Id of the script
- * @property {IContractEdge | IEdge} raw - Raw transaction object
- * @property {string} paymentId - Payment Id of the script
- * @property {number} timestamp - Timestamp of the script
- */
-class FairScript {
-  private readonly _owner: string;
-  private readonly _name: string;
-  private readonly _txid: string;
-  private readonly _rawTx: IContractEdge | IEdge;
-  private readonly _paymentId: string;
-  private readonly _timestamp: number;
-
-  constructor(tx: IContractEdge | IEdge) {
-    const txid = findTag(tx, 'scriptTransaction');
-    if (!txid) {
-      throw new Error('Invalid script transaction');
-    }
-
-    this._owner = getTxOwner(tx);
-    this._name = findTag(tx, 'scriptName') ?? 'Name Not available';
-    this._txid = txid;
-    this._rawTx = tx;
-    this._paymentId = tx.node.id;
-    this._timestamp = parseInt(findTag(tx, 'unixTime') as string, 10);
-  }
-
-  public get owner() {
-    return this._owner;
-  }
-
-  public get name() {
-    return this._name;
-  }
-
-  public get txid() {
-    return this._txid;
-  }
-
-  public get raw() {
-    return this._rawTx;
-  }
-
-  public get paymentId() {
-    return this._paymentId;
-  }
-
-  public get timestamp() {
-    return this._timestamp;
-  }
-}
+} from '../utils';
 
 const commonTags: ITagFilter[] = [...DEFAULT_TAGS_RETRO, ...SCRIPT_CREATION_PAYMENT_TAGS];
 
@@ -184,7 +127,7 @@ function listScripts(modelId?: string): Promise<FairScript[]>;
  */
 function listScripts(modelTx: IContractEdge | IEdge): Promise<FairScript[]>;
 
-function listScripts(param?: listScriptsParam) {
+function listScripts(param?: listFilterParams) {
   if (!param) {
     return _listAllScripts();
   } else if (param instanceof Object) {
