+++ src/types/arweave.ts
@@ -84,13 +84,4 @@ export interface IContractQueryResult {
   transactions: IContractTransactions;
 }
 
-export interface UState {
-  state: {
-    name: string;
-    ticker: string;
-    settings: Array<Array<string>>;
-    balances: { [address: string]: string };
-    claimable: Array<{ txid: string; to: string }>;
-    divisibility: number;
-  };
-}
+export type listFilterParams = string | IContractEdge | IEdge;
