+++ src/types/u.ts
@@ -1,0 +1,10 @@
+export interface UState {
+  state: {
+    name: string;
+    ticker: string;
+    settings: Array<Array<string>>;
+    balances: { [address: string]: string };
+    claimable: Array<{ txid: string; to: string }>;
+    divisibility: number;
+  };
+}
